Refactoring code in ContentScreen.jsx andin style.js
